Revert "Move "tracks" analytics into a shared calypso-analytics package (#39001)"
This reverts commit 2e3c19ab3711ef1d64fef7b632a66e3af34d4bb7.
