+++ client/lib/analytics/ad-tracking.js
@@ -15,16 +15,12 @@ import { loadScript } from '@automattic/load-script';
 import {
 	isPiiUrl,
 	costToUSD,
+	doNotTrack,
+	getCurrentUser,
 	isAdTrackingAllowed,
 	mayWeTrackCurrentUserGdpr,
 	refreshCountryCodeCookieGdpr,
 } from 'lib/analytics/utils';
-
-import {
-	getTracksAnonymousUserId,
-	getCurrentUser,
-	getDoNotTrack,
-} from '@automattic/calypso-analytics';
 
 /**
  * Module variables
@@ -1446,7 +1442,7 @@ function floodlightSessionId() {
 function floodlightUserParams() {
 	const params = {};
 	const currentUser = getCurrentUser();
-	const anonymousUserId = getTracksAnonymousUserId();
+	const anonymousUserId = tracksAnonymousUserId();
 
 	if ( currentUser ) {
 		params.u4 = currentUser.hashedPii.ID;
@@ -1711,7 +1707,7 @@ function setupWpcomFloodlightGtag() {
  * 4. the current user could be in the GDPR zone and hasn't consented to tracking
  * 5. `document.location.href` may contain personally identifiable information
  *
- * Note that getDoNotTrack() and isPiiUrl() can change at any time which is why we do not cache them.
+ * Note that doNotTrack() and isPiiUrl() can change at any time which is why we do not cache them.
  *
  * @returns {boolean} true if GA is allowed.
  */
@@ -1719,7 +1715,7 @@ export function isGoogleAnalyticsAllowed() {
 	return (
 		isGoogleAnalyticsEnabled &&
 		config.isEnabled( 'ad-tracking' ) &&
-		! getDoNotTrack() &&
+		! doNotTrack() &&
 		! isPiiUrl() &&
 		mayWeTrackCurrentUserGdpr()
 	);
