+++ client/lib/analytics/hotjar.js
@@ -7,8 +7,7 @@ import debug from 'debug';
  * Internal dependencies
  */
 import config from 'config';
-import { mayWeTrackCurrentUserGdpr, isPiiUrl } from './utils';
-import { getDoNotTrack } from '@automattic/calypso-analytics';
+import { mayWeTrackCurrentUserGdpr, doNotTrack, isPiiUrl } from './utils';
 import { isE2ETest } from 'lib/e2e';
 
 const hotjarDebug = debug( 'calypso:analytics:hotjar' );
@@ -20,7 +19,7 @@ export function addHotJarScript() {
 		hotJarScriptLoaded ||
 		! config( 'hotjar_enabled' ) ||
 		isE2ETest() ||
-		getDoNotTrack() ||
+		doNotTrack() ||
 		isPiiUrl() ||
 		! mayWeTrackCurrentUserGdpr()
 	) {
