+++ client/lib/analytics/utils/index.js
@@ -3,6 +3,7 @@
  */
 export { default as MARKETING_COUPONS_KEY } from './marketing-coupons-key';
 export { default as costToUSD } from './cost-to-usd';
+export { default as doNotTrack } from './do-not-track';
 export { default as hashPii } from './hash-pii';
 export { default as isPiiUrl } from './is-pii-url';
 export { default as isUrlBlacklistedForPerformance } from './is-url-blacklisted-for-performance';
@@ -13,3 +14,4 @@ export { default as urlParseAmpCompatible } from './url-parse-amp-compatible';
 export { default as shouldReportOmitBlogId } from './should-report-omit-blog-id';
 export { default as saveCouponQueryArgument } from './save-coupon-query-argument';
 export { default as refreshCountryCodeCookieGdpr } from './refresh-country-code-cookie-gdpr';
+export { getCurrentUser, setCurrentUser } from './current-user';
