+++ client/lib/analytics/utils/is-ad-tracking-allowed.js
@@ -3,10 +3,11 @@
  */
 import config from 'config';
 import debug from './debug';
+import doNotTrack from './do-not-track';
 import isUrlBlacklistedForPerformance from './is-url-blacklisted-for-performance';
 import isPiiUrl from './is-pii-url';
 import mayWeTrackCurrentUserGdpr from './may-we-track-current-user-gdpr';
-import { getDoNotTrack } from '@automattic/calypso-analytics';
+
 /**
  * Returns whether ad tracking is allowed.
  *
@@ -22,7 +23,7 @@ import { getDoNotTrack } from '@automattic/calypso-analytics';
 export default function isAdTrackingAllowed() {
 	const result =
 		config.isEnabled( 'ad-tracking' ) &&
-		! getDoNotTrack() &&
+		! doNotTrack() &&
 		! isUrlBlacklistedForPerformance() &&
 		! isPiiUrl() &&
 		mayWeTrackCurrentUserGdpr();
