--- packages/calypso-analytics/jest.config.js
@@ -1,1 +1,0 @@
-module.exports = { preset: '@automattic/calypso-build', rootDir: __dirname };
